+++ src/index.ts
@@ -37,6 +37,7 @@ async function validatePassword(
   // Blocklist validation
   if (options.blocklist) {
     const blocklistResult = blocklistValidator(password, options.blocklist, {
+      trimWhitespace: options.trimWhitespace,
       matchingSensitivity: options.matchingSensitivity,
       minEditDistance: options.minEditDistance,
       maxEditDistance: options.maxEditDistance,